refactoring: loading data into cluster state
